Base class for companion radios
